# lib/api-zod/src/generated/types/bookingStatusUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sip of Sky Restaurant API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatusUpdateStatus } from './bookingStatusUpdateStatus';

export interface BookingStatusUpdate {
  status: BookingStatusUpdateStatus;
}