# lib/api-zod/src/generated/types/menuItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sip of Sky Restaurant API
 * OpenAPI spec version: 0.1.0
 */

export interface MenuItem {
  id: number;
  name: string;
  description: string;
  price: number;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  rating: number;
  isVegetarian: boolean;
  isSpicy: boolean;
  isFeatured: boolean;
  available: boolean;
  createdAt: string;
}